refactor(editor): give links, search, and menu their own capability folders
Each capability's state, behavior, and Lexical binding now sit under one path,
following the checklist slice. Links was the most scattered: its plugins in
`plugins/`, its picker in `plugins/note-link/`, its helpers in `links/`, and its
node type in `runtime/`. Menu and note-controls move together because
`note-state.ts` is shared by exactly those two and nothing else.

`links` and `search` stop being top-level buckets, so two edges are restated
rather than introduced: the search plugin reaching the Note SDK adapter, and the
view provider reading search candidates. Both were legal before under the names
those buckets had.

`plugins/` is left holding core editing, folding, and infrastructure — the
subject of the next two changes.
