+++ config/eslint/editorBoundaries.ts
@@ -17,8 +17,6 @@ export const boundariesElements = [
   { type: 'runtime', pattern: `${EDITOR}/runtime`, stopMatching: true },
   { type: 'features', pattern: `${EDITOR}/features`, stopMatching: true },
   { type: 'plugins', pattern: `${EDITOR}/plugins`, stopMatching: true },
-  { type: 'links', pattern: `${EDITOR}/links`, stopMatching: true },
-  { type: 'search', pattern: `${EDITOR}/search`, stopMatching: true },
   { type: 'triggers', pattern: `${EDITOR}/triggers`, stopMatching: true },
   { type: 'view', pattern: `${EDITOR}/view`, stopMatching: true },
   { type: 'adapters', pattern: `${EDITOR}/note-sdk-adapters`, stopMatching: true },
@@ -33,18 +31,14 @@ export const boundariesPolicies = [
     allow: { to: { element: { type: ['editor-root', 'runtime'] } } } },
   { from: { element: { type: 'runtime' } },
     allow: { to: { element: { type: ['editor-root', 'outline', 'features'] } } } },
-  { from: { element: { type: 'links' } },
-    allow: { to: { element: { type: ['outline', 'runtime'] } } } },
-  { from: { element: { type: 'search' } },
-    allow: { to: { element: { type: ['outline'] } } } },
   { from: { element: { type: 'triggers' } },
     allow: { to: { element: { type: ['outline', 'runtime'] } } } },
   { from: { element: { type: 'view' } },
-    allow: { to: { element: { type: ['outline', 'search'] } } } },
+    allow: { to: { element: { type: ['outline', 'features'] } } } },
   { from: { element: { type: 'features' } },
-    allow: { to: { element: { type: ['editor-root', 'outline', 'runtime', 'triggers', 'view'] } } } },
+    allow: { to: { element: { type: ['editor-root', 'outline', 'runtime', 'triggers', 'view', 'adapters'] } } } },
   { from: { element: { type: 'plugins' } },
-    allow: { to: { element: { type: ['editor-root', 'features', 'links', 'adapters', 'outline', 'runtime', 'triggers', 'view'] } } } },
+    allow: { to: { element: { type: ['editor-root', 'features', 'adapters', 'outline', 'runtime', 'triggers', 'view'] } } } },
   { from: { element: { type: 'editor-root' } },
     allow: { to: { element: { type: ['features', 'outline', 'plugins', 'runtime'] } } } },
   { from: { element: { type: 'adapters' } },
