+++ src/client/editor/Editor.tsx
@@ -15,7 +15,7 @@ import { InsertionPlugin } from './plugins/InsertionPlugin';
 import { DeletionPlugin } from './plugins/DeletionPlugin';
 import { DatePlugin } from './features/date/DatePlugin';
 import { NoteBodyPlugin } from './features/note-body/NoteBodyPlugin';
-import { LinksPlugin } from './plugins/LinksPlugin';
+import { LinksPlugin } from './features/links/LinksPlugin';
 import { RootSchemaPlugin } from './plugins/RootSchemaPlugin';
 import { NoteIdPlugin } from './plugins/NoteIdPlugin';
 import { ClipboardPlugin } from './plugins/ClipboardPlugin';
@@ -25,10 +25,10 @@ import { StatusIndicators } from './StatusIndicators';
 import { ZoomPlugin } from './features/zoom/ZoomPlugin';
 import { ZoomVisibilityPlugin } from './features/zoom/ZoomVisibilityPlugin';
 import { FoldingPlugin } from './plugins/FoldingPlugin';
-import { NoteControlsPlugin } from './plugins/NoteControlsPlugin';
-import { NoteMenuPlugin } from './plugins/NoteMenuPlugin';
+import { NoteControlsPlugin } from './features/menu/NoteControlsPlugin';
+import { NoteMenuPlugin } from './features/menu/NoteMenuPlugin';
 import { MobileActionToolbar } from './plugins/mobile-toolbar/MobileActionToolbar';
-import { SearchCandidatesPlugin } from './plugins/SearchCandidatesPlugin';
+import { SearchCandidatesPlugin } from './features/search/SearchCandidatesPlugin';
 import { PendingDocumentImportPlugin } from './plugins/PendingDocumentImportPlugin';
 import './Editor.css';
