+++ src/client/editor/features/links/BlankTargetLinkInterceptorPlugin.tsx
