+++ src/client/editor/features/links/note-link-node.ts
