+++ src/client/editor/features/menu/NoteControlsPlugin.tsx
@@ -7,7 +7,7 @@ import { createPortal } from 'react-dom';
 import { OPEN_NOTE_MENU_COMMAND, SET_NOTE_FOLD_COMMAND } from '#client/editor/commands';
 import { resolveContentItemFromNode } from '#client/editor/outline/schema';
 import { focusEditorRoot } from '#client/editor/runtime/focus';
-import { $resolveNoteStateFromDOMNode } from '#client/editor/plugins/note-state';
+import { $resolveNoteStateFromDOMNode } from '#client/editor/features/menu/note-state';
 import { useZoomNoteId } from '#client/editor/view/EditorViewProvider';
 
 interface NoteControlsState {
