+++ src/client/editor/features/menu/NoteMenuPlugin.tsx
@@ -18,9 +18,9 @@ import { focusEditorRoot } from '#client/editor/runtime/focus';
 import { requireContentItemFromNode } from '#client/editor/outline/schema';
 import { installOutlineSelectionHelpers } from '#client/editor/outline/selection/store';
 import { getNestedList } from '#client/editor/outline/selection/tree';
-import { handleNoteMenuShortcut } from '#client/editor/plugins/note-menu-shortcuts';
-import type { NoteMenuShortcutEvent } from '#client/editor/plugins/note-menu-shortcuts';
-import { $resolveNoteStateFromDOMNode } from '#client/editor/plugins/note-state';
+import { handleNoteMenuShortcut } from '#client/editor/features/menu/note-menu-shortcuts';
+import type { NoteMenuShortcutEvent } from '#client/editor/features/menu/note-menu-shortcuts';
+import { $resolveNoteStateFromDOMNode } from '#client/editor/features/menu/note-state';
 import { useZoomNoteId } from '#client/editor/view/EditorViewProvider';
 import { isOtherPopupActive, setPopupActive } from '#client/editor/triggers/active-popup';
