+++ src/client/editor/features/menu/note-menu-shortcuts.ts
