+++ src/client/editor/features/menu/note-state.ts
