+++ src/client/editor/features/search/SearchCandidatesPlugin.tsx
