+++ src/client/editor/plugins/clipboard-note-ids.spec.ts
@@ -3,7 +3,7 @@ import { $isLineBreakNode } from 'lexical';
 import { waitFor } from '@testing-library/react';
 
 import type { SerializedLexicalNode, SerializedTextNode } from 'lexical';
-import type { SerializedNoteLinkNode } from '#client/editor/runtime/note-link-node';
+import type { SerializedNoteLinkNode } from '#client/editor/features/links/note-link-node';
 import type { SerializedNoteListItemNode } from '#client/editor/runtime/serialized-note-types';
 import type { RemdoTestApi } from '#client/editor/plugins/dev';
 import { collectOutlineNoteIds, flattenOutline } from '#tests-common/outline';
