+++ tests/unit/collab/snapshot.collab.spec.ts
@@ -8,7 +8,7 @@ import { meta } from '#tests';
 import { afterEach, describe, expect, it } from 'vitest';
 import { waitFor } from '@testing-library/react';
 import type { Buffer } from 'node:buffer';
-import { $isNoteLinkNode } from '#client/editor/runtime/note-link-node';
+import { $isNoteLinkNode } from '#client/editor/features/links/note-link-node';
 import { stripEditorStateDefaults } from '#tools/editor-state-defaults';
 import { $findNoteById } from '#client/editor/outline/note-traversal';
 import { COLLAB_LONG_TIMEOUT_MS } from './_support/timeouts';
