+++ tests/unit/note-link-node.spec.ts
@@ -1,7 +1,7 @@
 import { $createTextNode } from 'lexical';
 import { describe, expect, it } from 'vitest';
 
-import { $createNoteLinkNode, $isNoteLinkNode, NoteLinkNode } from '#client/editor/runtime/note-link-node';
+import { $createNoteLinkNode, $isNoteLinkNode, NoteLinkNode } from '#client/editor/features/links/note-link-node';
 import { meta } from '#tests';
 import { $findNoteById } from '#client/editor/outline/note-traversal';
